Force utf-8 encoding on all files
We really should guess source encoding and allow user to specify
destination encoding, but this is easy to get wrong. It seems that
Pelican supports only UTF-8 anyway, so we shouldn't feel too bad about
going the easy way here.

Oh, this is actually needed to read UTF-8 files on Windows. Most
Linuxes should not need it anyway.